Add files via upload
Still trying to find a GUI that looks decent.
Made the dealer and gave it game-specific functions like not drawing if it has a score above 17 and ending the game if it has a score over 21.
Started working debugging a GUI bug wherein the Gui_Start.py when the user draws a new card the image of the new card will pop up but the game will not keep the last card drawn on the screen. Believe it is an issue with the ordering of actions taken when the 'draw' button is pressed.